Factor out task actions into ds_task.[ch]
